ui: remove account screen
